move http stuff to 'client.rs'
